Initial work on XD-build data model implementation
The basic idea is to extend the OE-lite metadata model with types.
Instead of only supporting strings, add proper types for lists, dicts,
integers, booleans and so on, and enforce specific types for given
variables.

The Python API for working with these variables are also refined to make
it much more convenient to work with, and hopefully making it less painful
to switch between recipe and python syntax.

Signed-off-by: Esben Haabendal <esben@haabendal.dk>